# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Courses API",
        "description": "API documentation for Courses",
        "version": "1.0.0"
    },
    "paths": {
        "/api/courses": {
            "get": {
                "summary": "Get list of courses",
                "operationId": "02eeda67042dba2d22a517f26fb8eaf1",
                "responses": {
                    "200": {
                        "description": "A list of courses"
                    }
                }
            },
            "post": {
                "summary": "Create a new course",
                "operationId": "a4d6b7e7be59571034428ffbb86aaf0a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Course Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Course Description"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Published",
                                            "Pending"
                                        ]
                                    },
                                    "is_premium": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/courses/{id}": {
            "get": {
                "summary": "Get a course by ID",
                "operationId": "f170bd35e3106c9d9bf5807499c80b68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A course"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            },
            "put": {
                "summary": "Update a course",
                "operationId": "4d00e54a4b57b3259c40894ad2d0c76b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Course Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated Course Description"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Published",
                                            "Pending"
                                        ]
                                    },
                                    "is_premium": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a course",
                "operationId": "50a6b8ff1c84c83d6049f8b067438cc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course deleted successfully"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            }
        }
    }
}